requirements.txt
# Valid for Python 3.6.3 and Python 2.7.14
# For python 2.7
pysha3
mock

# Core
argon2_cffi
kivy
kivy.deps.sdl2
kivy.deps.glew

# Tests
nose
rednose
coverage
coveralls